+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Mighty Change Log
 
+## 3.1.2
+
+- Convert DataContractStore and TableMetaDataStore from Dictionary to ConcurrentDictionary
+- Fix bug in DataContractKey.Equals
+
 ## 3.1.1
 
 - Minor updates to NuGet package description only